move logs to own file
